move otlp types into otlp.rs
